Moves views and view models into "v1" directory
This will make API versioning much easier.

The idea is that URLs with no version will be "current", but there
will also be a set of versioned URLs that won't change.

Minor version changes will be backwards compatible within the current major version.
Major version changes may break backwards compatability.
