--- cegs_portal/search/view_models/__init__.py
@@ -1,5 +1,0 @@
-from .dhss import DHSSearch
-from .experiment import ExperimentSearch
-from .features import FeatureSearch, IdType
-from .reg_effects import RegEffectSearch
-from .search import Search
