+++ cegs_portal/search/view_models/v1/reg_effects.py
