+++ cegs_portal/search/views/__init__.py
@@ -1,6 +1,2 @@
-from .dhss import DHS, DHSLoc
-from .experiment import ExperimentView
-from .features import Feature, FeatureEnsembl, FeatureLoc
+from . import v1
 from .index import index
-from .reg_effects import RegEffectView
-from .search import SearchView
