+++ cegs_portal/search/views/v1/__init__.py
@@ -1,0 +1,5 @@
+from .dhss import DHS, DHSLoc
+from .experiment import ExperimentView
+from .features import Feature, FeatureEnsembl, FeatureLoc
+from .reg_effects import RegEffectView
+from .search import SearchView
