+++ cegs_portal/search/views/v1/features.py
@@ -1,4 +1,4 @@
-from cegs_portal.search.view_models import FeatureSearch, IdType
+from cegs_portal.search.view_models.v1 import FeatureSearch, IdType
 from cegs_portal.search.views.custom_views import TemplateJsonView
 from cegs_portal.search.views.renderers import json
