+++ cegs_portal/search/views/v1/reg_effects.py
@@ -1,4 +1,4 @@
-from cegs_portal.search.view_models import RegEffectSearch
+from cegs_portal.search.view_models.v1 import RegEffectSearch
 from cegs_portal.search.views.custom_views import TemplateJsonView
